pyproject.toml
[build-system]
requires = ["setuptools>=75.6.0", "wheel>=0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "bezzanlabs.gpt_docgen"
dynamic = ["version"]
authors = [
  {name="Vitor Bezzan", email="vitor@bezzan.com"},
]
description = "CLI tool to generate project documentation using LLMs."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb~=0.5.3",
    "langchain~=0.2.6",
    "langchain_chroma~=0.1.2",
    "langchain_community~=0.2.6",
    "langchain_openai",
    "langchain-anthropic",
    "openai~=1.35.7",
    "pydantic==2.7.4",
    "python-dotenv",
    "typer~=0.12.3",
]

[tool.setuptools.dynamic]
version = {attr = "constants.__version__"}

[project.optional-dependencies]
dev = [
    "pre-commit==3.7.1",
    "mypy==1.10.1",
    "pytest==8.2.2",
    "pytest-mock",
    "pytest-cov",
    "pandas-stubs==2.2.2.240603",
    "mkdocs==1.6.1",
    "mkdocs-material==9.5.44",
    "mkdocstrings-python==1.12.2",
]

[project.urls]
"Homepage" = "https://github.com/vitorbezzan/gpt_docgen"
"Bug Tracker" = "https://github.com/vitorbezzan/gpt_docgen"

[project.scripts]
gpt_docgen = "__init__:start_application"

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report html --cov-report term-missing --cov-fail-under 0"
pythonpath = [
  "src"
]
